+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * `API`:
     * All `with*Action` extensions got a contracts which declare that `block` will be called once
+    * Add several extensions `TelegramBot#sendPhoto` with `PhotoSize`
+    * Add several extensions `TelegramBot#reply` with `PhotoSize`
 
 ## 0.38.9